+++ PSMermaid.psm1
@@ -1,3 +1,8 @@
+# using smartAliases
+foreach($EntryType in @('Commit',"Branch",'Merge','Checkout','Cherrypick')){
+    New-Alias -Name "New-MermaidGitGraphEntry$EntryType" -Value New-MermaidGitGraphEntry
+}
+
 $FunctionFiles = $("$PSScriptRoot\Public\","$PSScriptRoot\Private\")| Get-Childitem -file -Recurse -Include "*.ps1" -ErrorAction SilentlyContinue
 
 foreach($FunctionFile in $FunctionFiles){
@@ -9,4 +14,4 @@ foreach($FunctionFile in $FunctionFiles){
     }
 }
 
-Export-ModuleMember -Function $FunctionFiles.BaseName
+Export-ModuleMember -Function $FunctionFiles.BaseName -Alias *